port ORCHdfsFileStream to OrcScanner
